+++ util/analysis/iplDepth.m
@@ -46,9 +46,6 @@ function [iplPercent, stats] = iplDepth(Neuron, varargin)
 	assert(isa(Neuron, 'sbfsem.core.StructureAPI'),...
 		'Input a StructureAPI object');
     
-    % GCL = sbfsem.builtin.GCLBoundary(Neuron.source, true);
-    % INL = sbfsem.builtin.INLBoundary(Neuron.source, true);
-    
     ip = inputParser();
     ip.CaseSensitive = false;
     addParameter(ip, 'numBins', 20, @isnumeric);
@@ -81,17 +78,6 @@ function [iplPercent, stats] = iplDepth(Neuron, varargin)
         nodes(nodes.Rum > 0.8*somaRadius, :) = [];
     end
 
-%	xyz = nodes.XYZum;
-% 
-% 	[X, Y] = meshgrid(GCL.newXPts, GCL.newYPts);
-% 	vGCL = interp2(X, Y, GCL.interpolatedSurface,...
-% 		xyz(:,1), xyz(:,2));
-% 
-% 	[X, Y] = meshgrid(INL.newXPts, INL.newYPts);
-% 	vINL = interp2(X, Y, INL.interpolatedSurface,...
-% 		xyz(:,1), xyz(:,2));
-% 
-% 	iplPercent = (xyz(:, 3) - vINL) ./ ((vGCL - vINL)+eps);
     iplPercent = micron2ipl(nodes.XYZum, Neuron.source);
 	iplPercent(isnan(iplPercent)) = [];
     
@@ -173,8 +159,10 @@ function [iplPercent, stats] = iplDepth(Neuron, varargin)
     
     if ip.Results.omitOutliers
         xlim(ax, [-0.25, 1.25]);
-        set(ax, 'XTick', 0:0.25:1,... 
-        'XTickLabel', {'INL', 'off', 'IPL', 'on', 'GCL'});
+        set(ax, 'XTick', 0:0.2:1,...
+            'XTickLabel', {'INL', '20', '40', '60', '80', 'GCL'});
+        %set(ax, 'XTick', 0:0.25:1,... 
+        %'XTickLabel', {'INL', 'off', 'IPL', 'on', 'GCL'});
     else
         xBound = [-0.25, 1.25];
         if min(iplPercent) < xBound(1)