+++ util/plots/hideAxes.m
@@ -1,8 +1,8 @@
 function hideAxes(gObj)
-    % HIDEAXES
+    % TOGGLEAXES
     % 
     % Description:
-    %   Toggle axes visibility without removing completely
+    %   Hide axes visibility without removing completely
     % 
     % Syntax:
     %   hideAxes(gObj);
@@ -10,46 +10,34 @@ function hideAxes(gObj)
     % Inputs:
     %   gObj      Figure or axes handle
     %
+    % See also:
+    %   TOGGLEAXES
+    %
     % History:
     %   6Jan2018 - SSP
     %   24Jan2018 - SSP - added option to show axes again
     %   25Apr2018 - SSP - applies to all axes in a figure
+    %   4Feb2020 - SSP - removed toggle options
     % ---------------------------------------------------------------------
+
     if nargin < 1
         gObj = gcf;
     else
-        assert(ishandle(gObj), 'Input an axes or figure handle');
+        assert(ishandle(gObj), 'Input axes or figure handle');
     end
-    
+
     switch class(gObj)
         case 'matlab.graphics.axis.Axes'
             ax = gObj;
         case 'matlab.ui.Figure'
             ax = findall(gObj, 'Type', 'axes');
         otherwise
-            warning('hideAxes:InvalidInput',...
-                'Input must be handle to a figure or axes');
-            return
+            warning('toggleAxes:InvalidInput',...
+                'Input must be handle to a figure or axes')
     end
-    
+
     for i = 1:numel(ax)
         bkgd = ax(i).Color;
-        frgd = invertColor(bkgd);
-        
-        if ax(i).XColor == bkgd
-            set(ax(i), 'XColor', frgd, 'YColor', frgd', 'ZColor', frgd); 
-        else
-            set(ax(i), 'XColor', bkgd, 'YColor', bkgd, 'ZColor', bkgd);
-        end
+        set(ax(i), 'XColor', bkgd, 'YColor', bkgd, 'ZColor', bkgd);
     end
-end
 
-function newColor = invertColor(oldColor)
-    if oldColor == [0, 0, 0]
-        newColor = 'w';
-    elseif oldColor == [1, 1, 1]
-        newColor = 'k';
-    else % Do nothing if not black/white
-        newColor = oldColor;
-    end
-end